Show calendar picker on Calendar Sync upgrade
