+++ app/src/main/java/io/ipoli/android/store/fragments/UpgradeStoreFragment.java
@@ -1,8 +1,11 @@
 package io.ipoli.android.store.fragments;
 
+import android.Manifest;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.app.LoaderManager;
+import android.support.v4.content.Loader;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -16,32 +19,46 @@ import com.squareup.otto.Subscribe;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import javax.inject.Inject;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
+import io.ipoli.android.Constants;
 import io.ipoli.android.R;
+import io.ipoli.android.app.AndroidCalendarEventParser;
 import io.ipoli.android.app.App;
 import io.ipoli.android.app.BaseFragment;
+import io.ipoli.android.app.SyncAndroidCalendarProvider;
 import io.ipoli.android.app.events.EventSource;
 import io.ipoli.android.app.events.ScreenShownEvent;
+import io.ipoli.android.app.events.SyncCalendarRequestEvent;
+import io.ipoli.android.app.persistence.CalendarPersistenceService;
+import io.ipoli.android.app.sync.AndroidCalendarLoader;
+import io.ipoli.android.app.ui.dialogs.AndroidCalendarsPickerFragment;
+import io.ipoli.android.app.ui.dialogs.LoadingDialog;
 import io.ipoli.android.app.utils.DateUtils;
 import io.ipoli.android.player.UpgradeManager;
+import io.ipoli.android.quest.data.Category;
 import io.ipoli.android.store.Upgrade;
 import io.ipoli.android.store.activities.StoreActivity;
 import io.ipoli.android.store.adapters.UpgradeStoreAdapter;
 import io.ipoli.android.store.events.BuyUpgradeEvent;
 import io.ipoli.android.store.events.UpgradeUnlockedEvent;
 import io.ipoli.android.store.viewmodels.UpgradeViewModel;
+import pub.devrel.easypermissions.AfterPermissionGranted;
+import pub.devrel.easypermissions.EasyPermissions;
+
+import static io.ipoli.android.Constants.RC_CALENDAR_PERM;
 
 /**
  * Created by Polina Zhelyazkova <polina@ipoli.io>
  * on 5/23/17.
  */
 
-public class UpgradeStoreFragment extends BaseFragment {
+public class UpgradeStoreFragment extends BaseFragment implements LoaderManager.LoaderCallbacks<Void> {
 
     @Inject
     Bus eventBus;
@@ -49,11 +66,22 @@ public class UpgradeStoreFragment extends BaseFragment {
     @Inject
     UpgradeManager upgradeManager;
 
+    @Inject
+    CalendarPersistenceService calendarPersistenceService;
+
+    @Inject
+    AndroidCalendarEventParser androidCalendarEventParser;
+
+    @Inject
+    SyncAndroidCalendarProvider syncAndroidCalendarProvider;
+
     private Unbinder unbinder;
 
     @BindView(R.id.upgrade_list)
     RecyclerView upgradeList;
     private UpgradeStoreAdapter adapter;
+    private LoadingDialog loadingDialog;
+    private Map<Long, Category> selectedCalendars;
 
 
     @Nullable
@@ -116,10 +144,37 @@ public class UpgradeStoreFragment extends BaseFragment {
             message = getString(R.string.upgrade_successfully_bought, title);
             adapter.setViewModels(createUpgradeViewModels());
             eventBus.post(new UpgradeUnlockedEvent(upgrade));
+            if (upgrade == Upgrade.CALENDAR_SYNC) {
+                pickCalendarsToSync();
+            }
+
         } else {
             message = getString(R.string.upgrade_too_expensive, title);
         }
         Toast.makeText(getContext(), message, Toast.LENGTH_SHORT).show();
+    }
+
+    private void pickCalendarsToSync() {
+        if (EasyPermissions.hasPermissions(getContext(), Manifest.permission.READ_CALENDAR)) {
+            showCalendarsPicker();
+        } else {
+            EasyPermissions.requestPermissions(this, getString(R.string.allow_read_calendars_perm_reason), RC_CALENDAR_PERM, Manifest.permission.READ_CALENDAR);
+        }
+    }
+
+    @AfterPermissionGranted(Constants.RC_CALENDAR_PERM)
+    private void showCalendarsPicker() {
+        AndroidCalendarsPickerFragment.newInstance(R.string.choose_calendars_title, getPlayer().getAndroidCalendars(), this::onSelectCalendarsToSync)
+                .show(getFragmentManager());
+    }
+
+    private void onSelectCalendarsToSync(Map<Long, Category> selectedCalendars) {
+        loadingDialog = LoadingDialog.show(getContext(), getString(R.string.sync_calendars_loading_dialog_title), getString(R.string.sync_calendars_loading_dialog_message));
+        this.selectedCalendars = selectedCalendars;
+        if (!selectedCalendars.isEmpty()) {
+            eventBus.post(new SyncCalendarRequestEvent(selectedCalendars, EventSource.STORE_UPGRADES));
+        }
+        getLoaderManager().initLoader(1, null, this);
     }
 
     @Override
@@ -142,5 +197,30 @@ public class UpgradeStoreFragment extends BaseFragment {
     @Override
     protected boolean useOptionsMenu() {
         return false;
+    }
+
+    @Override
+    public Loader<Void> onCreateLoader(int id, Bundle args) {
+        return new AndroidCalendarLoader(getContext(), selectedCalendars, getPlayer(), syncAndroidCalendarProvider, androidCalendarEventParser, calendarPersistenceService);
+    }
+
+    @Override
+    public void onLoadFinished(Loader<Void> loader, Void data) {
+        if (loadingDialog != null) {
+            loadingDialog.dismiss();
+            loadingDialog = null;
+        }
+    }
+
+    @Override
+    public void onLoaderReset(Loader<Void> loader) {
+        // intentional
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        EasyPermissions.onRequestPermissionsResult(requestCode, permissions, grantResults, this);
     }
 }
